Bug 805917 - Add x86-64 support. r=jandem
UltraBlame original commit: 4d42c3ef72f58380264349fa5dbe4b276be9750d

+++ js/src/ion/shared/IonFrames-x86-shared.h
@@ -442,7 +442,7 @@ class InvalidationBailoutStack
 
 class BaselineFrame
 {
-    uint32_t frameSize;
+    size_t frameSize;
 
   public:
